Made io_result in pool.hpp be an int64_t.

+++ src/arch/io/disk/pool.hpp
@@ -55,9 +55,7 @@ public:
         size_t count;
         off_t offset;
 
-        // TODO: io_result almost probably definitely should be an
-        // int64_t or was at some point, like in libaio.
-        int io_result;
+        int64_t io_result;
 
         void run();
         void done();
